ref(explore): Extract findFreshEmptyPrefixSearchCacheMatch to shared utility
Move the prefix-search cache matching logic from traceItemAttributeKeysOptions
into its own module so it can be reused by other attribute value hooks.
